Change fdev1 Base To debian Instead Of gcc:7
We had an issue in our unit tests where the versions of libstdc++.so
between fdev1 and fdev2 were slightly different.  Switching the base
image to be the same between the fdev1 development environment and
dtest1 the test or production environment seems to eliminate the
differences.

gcc is based on Debian but it has a non package based instal of GCC on
it.  Debian with build-essentials should keep everything within the
packed tools realm which has hopefull been kept in synch by someone
else.