Make Common base Image
Try to reduce the size and build time of the Docker container by making
a common base image that works for both builder and the final Docker
Image.

This should give a single layer within the Docker iamge that is used in
both places, reducing space use and rebuild time.