Combine Multiple COPY Command Lines
Each COPY command line adds a new layer to the Docker Cache.  This can
increase  (re-)build times and will increase image size.  Combine copies
that go to the same target directory together.

In the Dockerfile COPY command the last parameter is the destination and
must be a Directory with a trailing / for multiple sources.